+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Gem dump x2
 - Dry configurable config (#20)
 - added .rspec for default spec helper require
+- Added SSL capabilities
 
 ## 0.3.1
 - Dev tools update